# swagger/openapi.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "API prueba técnica",
    "description": "Proyecto enfocado en realizar pruebas de codificación.",
    "termsOfService": "Terms of service",
    "contact": {
      "name": "Jorge Iraira",
      "url": "https://www.linkedin.com/in/jorge-iraira-ab099476/",
      "email": "jorgeiraira55@gmail.com"
    },
    "license": {
      "name": "License of API",
      "url": "API license URL"
    },
    "version": "1.0"
  },
  "servers": [
    {
      "url": "http://localhost:8080/",
      "description": "Local server"
    },
    {
      "url": "https://xx.qa.midominio.com/",
      "description": "QA server"
    },
    {
      "url": "https://xx.staging.midominio.com/",
      "description": "Staging server"
    },
    {
      "url": "https://xx.produccion.com/",
      "description": "Production server"
    }
  ],
  "paths": {
    "/prices": {
      "get": {
        "tags": [
          "price-controller"
        ],
        "summary": "Obtener información de precio a aplicar",
        "operationId": "getPrice",
        "parameters": [
          {
            "name": "applicationDate",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "format": "date-time"
            }
          },
          {
            "name": "productId",
            "in": "query",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "brandId",
            "in": "query",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResultResponse"
                },
                "examples": {
                  "Ejemplo de Price Response": {
                    "summary": "Ejemplo de respuesta correcta",
                    "description": "Ejemplo de Price Response",
                    "value": {
                      "status": 500,
                      "message": "An unexpected error occurred: Required request parameter 'brandId' for method parameter type Integer is not present",
                      "data": null
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "Price not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResultResponse"
                },
                "examples": {
                  "Ejemplo de Price Response": {
                    "summary": "Ejemplo de respuesta correcta",
                    "description": "Ejemplo de Price Response",
                    "value": {
                      "status": 404,
                      "message": "Price not found: Applicable price not found for productId: 35455, brandId: 2, and date: 2020-06-14T19:00:59",
                      "data": null
                    }
                  }
                }
              }
            }
          },
          "200": {
            "description": "Correcta obtención de price",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResultResponse"
                },
                "examples": {
                  "Ejemplo de Price Response": {
                    "summary": "Ejemplo de respuesta correcta",
                    "description": "Ejemplo de Price Response",
                    "value": {
                      "status": 200,
                      "message": "success",
                      "data": {
                        "priceList": 1,
                        "brandId": 1,
                        "startDate": "2020-06-14T00:00:00",
                        "endDate": "2020-12-31T23:59:59",
                        "productId": 35455,
                        "priority": 0,
                        "price": "35.50",
                        "curr": "EUR"
                      }
                    }
                  }
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "ApiResultResponse": {
        "type": "object",
        "properties": {
          "status": {
            "type": "integer",
            "format": "int32"
          },
          "message": {
            "type": "string"
          },
          "data": {
            "type": "object"
          }
        }
      }
    }
  }
}